% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictHPXCS.R
\name{predictHPXCS}
\alias{predictHPXCS}
\title{Title predictcluster}
\usage{
predictHPXCS(blcadata)
}
\arguments{
\item{blcadata}{blcadata is a gene expression matrix form rna-seq or array. rowname is gene name and colname is sample's name. The value need log2(data+1) process before predicted.}
}
\value{

}
\description{
Title predictcluster
}
\examples{
cluster <- predictHPXCS(test)
}
